refactor(search): components
