+++ src/lib/components/search/display/general/Result.svelte
