+++ src/lib/components/search/display/images/Image.svelte
