+++ src/lib/components/search/display/images/Preview.svelte
