+++ src/lib/components/search/load/general/LoadResult.svelte
@@ -1,0 +1,14 @@
+<div class="w-full animate-pulse">
+	<div
+		class="w-max-content min-[292px]:w-[292px] h-[20px] bg-gray-800 dark:bg-gray-400 rounded-full"
+	></div>
+	<div
+		class="my-0.5 w-max-content min-[230px]:w-[218px] h-[28px] bg-hearchco-primary dark:bg-hearchco-secondary rounded-lg"
+	></div>
+	<div
+		class="my-0.5 w-max-content sm:w-[620px] h-10 text-sm bg-gray-600 dark:bg-gray-200 rounded-lg"
+	></div>
+	<div class="my-2 flex justify-end rounded-lg">
+		<div class="w-[115px] h-[15px] bg-gray-800 dark:bg-gray-400 rounded-lg"></div>
+	</div>
+</div>
